+++ validator-api/src/main/java/no/difi/vefa/validator/api/DeclarationWithChildren.java
@@ -1,0 +1,7 @@
+package no.difi.vefa.validator.api;
+
+import java.io.InputStream;
+
+public interface DeclarationWithChildren extends Declaration {
+    Iterable<InputStream> children(InputStream inputStream);
+}