+++ validator-api/src/main/java/no/difi/vefa/validator/api/DeclarationWithConverter.java
@@ -3,6 +3,6 @@ package no.difi.vefa.validator.api;
 import java.io.InputStream;
 import java.io.OutputStream;
 
-public interface DeclarationAndConverter extends Declaration {
+public interface DeclarationWithConverter extends Declaration {
     void convert(InputStream inputStream, OutputStream outputStream) throws ValidatorException;
 }