+++ validator-api/src/main/java/no/difi/vefa/validator/api/ValidatorPlugin.java
@@ -1,0 +1,10 @@
+package no.difi.vefa.validator.api;
+
+import java.util.List;
+
+public interface ValidatorPlugin {
+    List<String> capabilities();
+    List<Class<? extends Checker>> checkers();
+    List<Declaration> declarations();
+    List<Class<? extends Renderer>> renderers();
+}