+++ validator-core/src/main/java/no/difi/vefa/validator/ValidatorInstance.java
@@ -16,6 +16,7 @@ import java.io.OutputStream;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 /**
  * Contains CheckerPools and Configuration, and is entry point for validation.
@@ -63,12 +64,12 @@ class ValidatorInstance implements Closeable {
      * @param source Source for validation artifacts
      * @throws ValidatorException
      */
-    ValidatorInstance(Source source, Properties properties, Class<? extends Checker>[] checkerImpls, Class<? extends Renderer>[] rendererImpls, Declaration[] declarations) throws ValidatorException {
+    ValidatorInstance(Source source, Properties properties, Class<? extends Checker>[] checkerImpls, Class<? extends Renderer>[] rendererImpls, Declaration[] declarations, Set<String> capabilities) throws ValidatorException {
         // Create config combined with default values.
         this.properties = new CombinedProperties(properties, ValidatorDefaults.PROPERTIES);
 
         // Create a new engine
-        validatorEngine = new ValidatorEngine(source.createInstance(this.properties));
+        validatorEngine = new ValidatorEngine(source.createInstance(this.properties, capabilities));
 
         // Declarations
         this.declarations = declarations;