+++ validator-core/src/main/java/no/difi/vefa/validator/declaration/UblDeclaration.java
@@ -8,12 +8,10 @@ import no.difi.vefa.validator.util.XmlUtils;
 
 import javax.xml.stream.XMLEventReader;
 import javax.xml.stream.XMLInputFactory;
-import javax.xml.stream.XMLStreamException;
-import javax.xml.stream.events.StartElement;
 import javax.xml.stream.events.Characters;
+import javax.xml.stream.events.StartElement;
 import javax.xml.stream.events.XMLEvent;
 import java.io.ByteArrayInputStream;
-import java.text.ParseException;
 
 /**
  * Document declaration for OASIS Universal Business Language (UBL).