+++ validator-core/src/main/java/no/difi/vefa/validator/source/RepositorySourceInstance.java
@@ -9,16 +9,14 @@ import org.slf4j.LoggerFactory;
 
 import javax.xml.bind.Unmarshaller;
 import java.net.URI;
+import java.util.Set;
 
 class RepositorySourceInstance extends AbstractSourceInstance {
 
     private static Logger logger = LoggerFactory.getLogger(RepositorySourceInstance.class);
 
-    private URI rootUri;
-
-    public RepositorySourceInstance(Properties properties, URI rootUri) throws ValidatorException {
+    public RepositorySourceInstance(Properties properties, Set<String> capabilities, URI rootUri) throws ValidatorException {
         super(properties);
-        this.rootUri = rootUri;
 
         try {
             Unmarshaller unmarshaller = jaxbContext.createUnmarshaller();
@@ -27,9 +25,18 @@ class RepositorySourceInstance extends AbstractSourceInstance {
             Artifacts artifactsType = (Artifacts) unmarshaller.unmarshal(artifactsUri.toURL());
 
             for (ArtifactType artifact : artifactsType.getArtifact()) {
-                URI artifactUri = rootUri.resolve(artifact.getFilename());
-                logger.info(String.format("Fetching %s", artifactUri));
-                unpackContainer(asicReaderFactory.open(artifactUri.toURL().openStream()), artifact.getFilename());
+                boolean loadArtifact = true;
+
+                if (artifact.getCapabilities() != null)
+                    for (String capability : artifact.getCapabilities().split(","))
+                        if (capabilities.contains(capability))
+                            loadArtifact = false;
+
+                if (loadArtifact) {
+                    URI artifactUri = rootUri.resolve(artifact.getFilename());
+                    logger.info(String.format("Fetching %s", artifactUri));
+                    unpackContainer(asicReaderFactory.open(artifactUri.toURL().openStream()), artifact.getFilename());
+                }
             }
         } catch (Exception e) {
             logger.warn(e.getMessage(), e);